+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### 🛑 Breaking changes 🛑
 
 - Remove deprecated pdata funcs/structs from v0.50.0 (#5345)
+- Remove deprecated pdata getters and setters of primitive slice values: `Value.BytesVal`, `Value.SetBytesVal`, 
+  `Value.UpdateBytes`, `Value.InsertBytes`, `Value.UpsertBytes`, `<HistogramDataPoint|Buckets>.BucketCounts`, 
+  `<HistogramDataPoint|Buckets>.SetBucketCounts`, `HistogramDataPoint.ExplicitBounds`,
+  `HistogramDataPoint.SetExplicitBounds` (#5347)
 
 ### 🚩 Deprecations 🚩